I Moved all methods that manage RawAthlete instances from RaceService class to RawAthleteService class
